move TestParseOID into identity_tests file
